# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "API documentation description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/body-types": {
            "get": {
                "tags": [
                    "Body Types"
                ],
                "summary": "Get a list of body types",
                "operationId": "dbd98e06b4df642cfc8ccf217b5e3600",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of body types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "SUV"
                                                    },
                                                    "icon": {
                                                        "type": "string",
                                                        "example": "suv.svg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Body Types"
                ],
                "summary": "Create a new body type",
                "operationId": "12d2320c2b87bcfb017aa258860207b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "icon"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sedan"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "sedan.svg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Body type created"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/admin/body-types/{id}": {
            "put": {
                "tags": [
                    "Body Types"
                ],
                "summary": "Update a body type",
                "operationId": "daea9d709612b138c7ea2d7b1f373e64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Body type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Coupe"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "coupe.svg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Body type updated"
                    },
                    "404": {
                        "description": "Body type not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Body Types"
                ],
                "summary": "Delete a body type",
                "operationId": "b2016acf607cf2bfd281ee164f88ca18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Body type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Body type deleted"
                    },
                    "404": {
                        "description": "Body type not found"
                    }
                }
            }
        },
        "/admin/body-types/bulk-destroy": {
            "post": {
                "tags": [
                    "Body Types"
                ],
                "summary": "Bulk delete body types",
                "operationId": "d34b4025e3dbfb80b5648f0314314757",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Body types deleted"
                    },
                    "422": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
...
}